Add creation of table for first run

+++ run_app.py
@@ -2,6 +2,7 @@ import os
 
 from things_organizer.main_app import app
 from things_organizer.data_management.user import delete_all_sessions
+from things_organizer.data_management import common
 
 from waitress import serve
 
@@ -10,9 +11,16 @@ config = {'ip': '127.0.0.1', 'port': 8080}
 
 # Generate a new Key each time the app is run.
 app.secret_key = os.urandom(16)
+
 # Comment the code below if you want to
 # delete sessions every time the server starts
-delete_all_sessions()
+if common.exists_database():
+    try:
+        delete_all_sessions()
+    except Exception as excerror:
+        pass
+else:
+    common.create_tables()
 
 # Run the server.
 # If you want a debugging server set
